fix: resolve lossy conversion from double to long in PDFUtil
- Adjusted calculation of total amount by rounding double to long using `Math.round()` to prevent lossy conversion issues.
- Added handling for precise rounding of currency values using `BigDecimal` for cases where more control over precision is required.
- Ensured consistent formatting of currency values (e.g., "Rp.") when adding cells to the summary table in PDF generation.
- Improved overall precision and type safety in financial calculations.

This fixes the issue related to incompatible types: possible lossy conversion from double to long.

+++ app/src/main/java/com/example/weighingscale/data/model/BatchDetail.java
@@ -19,7 +19,7 @@ public class BatchDetail {
     public String id;
     public String batch_id;
     public Date datetime;
-    public int amount;
+    public Double amount;
 
     // Constructor
     public BatchDetail() {
@@ -35,7 +35,7 @@ public class BatchDetail {
         return datetime;
     }
 
-    public int getAmount() {
+    public Double getAmount() {
         return amount;
     }
 }
